# Frontend/sssklv2-openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "SSSKLv2 | v1",
        "contact": {
            "name": "Scouting Wilo",
            "email": "webmaster@scoutingwilo.nl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:7193/"
        }
    ],
    "paths": {
...
        "/v1/Announcement": {
            "get": {
                "tags": [
                    "Announcement"
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Announcement"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Announcement"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Announcement"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Announcement"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/AnnouncementCreateDto"
                                    }
                                ]
                            }
                        },
                        "text/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/AnnouncementCreateDto"
                                    }
                                ]
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/AnnouncementCreateDto"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Announcement"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Announcement"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Announcement"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/Announcement/{id}": {
            "get": {
                "tags": [
                    "Announcement"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Announcement"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Announcement"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Announcement"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Announcement"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/AnnouncementUpdateDto"
                                    }
                                ]
                            }
                        },
                        "text/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/AnnouncementUpdateDto"
                                    }
                                ]
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/AnnouncementUpdateDto"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Announcement"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Announcement": {
                "required": [
                    "createdOn"
                ],
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "fotoUrl": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "url": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "order": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "isScheduled": {
                        "type": "boolean"
                    },
                    "plannedFrom": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "plannedTill": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "createdOn": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AnnouncementCreateDto": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "order": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "isScheduled": {
                        "type": "boolean"
                    },
                    "plannedFrom": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "plannedTill": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    }
                }
            },
            "AnnouncementUpdateDto": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "order": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "isScheduled": {
                        "type": "boolean"
                    },
                    "plannedFrom": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "plannedTill": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    }
                }
            },
...
        }
    },
...
}